- got rid of warnings on Mac OS - improved calculation of when GC should activate
git-svn-id: https://openmodelica.org/svn/MetaModelica/trunk@205 324bd646-64f9-0310-949b-d3e75d9ef174

+++ runtime/common/internal.c
@@ -63,7 +63,7 @@ void rmldb_var_print(void *p)
     if (!p) { printf ("NIL"); fflush(stdout); return; }
     if( RML_ISIMM(p) ) 
     {
-        printf ("%d", RML_UNTAGFIXNUM(p));    
+        printf ("%ld", (long)RML_UNTAGFIXNUM(p));    
     } 
     else 
     {
@@ -86,12 +86,12 @@ void rmldb_var_print(void *p)
                     void **pp = NULL;
                     if (slots == 0)
                     {
-                        printf ("{S(%d)[%d]=NIL}", constr, slots);
+		        printf ("{S(%ld)[%ld]=NIL}", (long)constr, (long)slots);
                         fflush(stdout);
                         return;
                     }
                     
-                    printf ("S(%d)[%d](", constr, slots);
+                    printf ("S(%ld)[%ld](", (long)constr, (long)slots);
 
                     pp = RML_STRUCTDATA(p);
                     fflush(stdout);
